vllm-test/requirements.txt
torch>=2.4.0
transformers>=4.48.0
accelerate>=0.34.0
peft>=0.12.0
trl>=0.13.0
bitsandbytes>=0.43.0
datasets>=3.0.0
pillow>=10.4.0
qwen_vl_utils
tqdm
bc
